+++ tensorcircuit/interfaces/__init__.py
@@ -3,5 +3,6 @@ Interfaces bridging different backends
 """
 
 from . import tensortrans
+from .tensortrans import numpy_args_to_backend, general_args_to_numpy
 from .scipy import scipy_interface, scipy_optimize_interface
 from .torch import torch_interface, pytorch_interface