+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 ## Unreleased
 
+### Added
+
+- PyTorch backend support multi pytrees version of ``tree_map``
+
 ### Changed
 
 - Refactor ``interfaces`` code as a submodule and add pytree support for args
 
 - Change the way to register global setup internally, so that we can skip the list of all submodules
+
+- Refactor the tensortrans code to a pytree perspective
 
 ## 0.2.1